refactor: 将原SpelMinValidator调整成抽象类，便于SpelMax的使用
